npm run prettier

+++ contracts/dao/governance/interfaces/IGovernor.sol
@@ -51,6 +51,7 @@ abstract contract IGovernor is IERC165 {
         bytes[] memory calldatas,
         bytes32 descriptionHash
     ) public virtual returns (uint256);
+
     function execute(
         address[] memory targets,
         uint256[] memory values,
@@ -60,7 +61,11 @@ abstract contract IGovernor is IERC165 {
 
     function castVote(uint256 proposalId, uint8 support) public virtual returns (uint256 balance);
 
-    function castVoteWithReason(uint256 proposalId, uint8 support, string memory reason) public virtual returns (uint256 balance);
+    function castVoteWithReason(
+        uint256 proposalId,
+        uint8 support,
+        string memory reason
+    ) public virtual returns (uint256 balance);
 
     function castVoteWithReasonAndParams(
         uint256 proposalId,
@@ -69,7 +74,13 @@ abstract contract IGovernor is IERC165 {
         bytes memory params
     ) public virtual returns (uint256 balance);
 
-    function castVoteBySig(uint256 proposalId, uint8 support, uint8 v, bytes32 r, bytes32 s) public virtual returns (uint256 balance);
+    function castVoteBySig(
+        uint256 proposalId,
+        uint8 support,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) public virtual returns (uint256 balance);
 
     function castVoteWithReasonAndParamsBySig(
         uint256 proposalId,
@@ -81,11 +92,19 @@ abstract contract IGovernor is IERC165 {
         bytes32 s
     ) public virtual returns (uint256 balance);
 
-    function getProposals(uint _numIndexes) public view virtual returns (string[] memory, string[] memory, string[] memory);
+    function getProposals(uint256 _numIndexes)
+        public
+        view
+        virtual
+        returns (
+            string[] memory,
+            string[] memory,
+            string[] memory
+        );
 
-    function getDescription(uint _proposalId) public view virtual returns (string memory);
+    function getDescription(uint256 _proposalId) public view virtual returns (string memory);
 
-    function getProposalIds() public view virtual returns (uint[] memory);
+    function getProposalIds() public view virtual returns (uint256[] memory);
 
     function name() public view virtual returns (string memory);
 
@@ -105,7 +124,11 @@ abstract contract IGovernor is IERC165 {
 
     function getVotes(address account, uint256 blockNumber) public view virtual returns (uint256);
 
-    function getVotesWithParams(address account, uint256 blockNumber, bytes memory params) public view virtual returns (uint256);
+    function getVotesWithParams(
+        address account,
+        uint256 blockNumber,
+        bytes memory params
+    ) public view virtual returns (uint256);
 
     function hasVoted(uint256 proposalId, address account) public view virtual returns (bool);
