+++ contracts/dao/tokens/MainToken.sol
@@ -6,7 +6,12 @@ pragma solidity 0.8.13;
 import "./ERC20/ERC20.sol";
 
 contract MainToken is ERC20 {
-    constructor(string memory name_, string memory symbol_, uint totalSupply_, address issuer) ERC20(name_, symbol_) {
+    constructor(
+        string memory name_,
+        string memory symbol_,
+        uint256 totalSupply_,
+        address issuer
+    ) ERC20(name_, symbol_) {
         _totalSupply = totalSupply_;
         _balances[issuer] = totalSupply_;
