+++ contracts/dao/treasury/MultiSigWallet.sol
@@ -13,23 +13,23 @@ contract MultiSigWallet is IMultiSigWallet {
         address to;
         bool executed;
         bytes data;
-        uint value;
-        uint numConfirmations;
-        uint expireTimestamp;
+        uint256 value;
+        uint256 numConfirmations;
+        uint256 expireTimestamp;
     }
 
     error TransactionRevered(bytes data);
 
-    uint public constant MAX_OWNER_COUNT = 50;
+    uint256 public constant MAX_OWNER_COUNT = 50;
 
     address[] public owners;
     address public governor;
 
-    uint public numConfirmationsRequired;
-    uint public lastDisabledTransactionIndex;
+    uint256 public numConfirmationsRequired;
+    uint256 public lastDisabledTransactionIndex;
 
     mapping(address => bool) public isOwner;
-    mapping(uint => mapping(address => bool)) public isConfirmed;
+    mapping(uint256 => mapping(address => bool)) public isConfirmed;
 
     mapping(address => bytes32) internal whitelistedBytesCode;
 
@@ -40,27 +40,27 @@ contract MultiSigWallet is IMultiSigWallet {
         _;
     }
 
-    modifier txExists(uint _txIndex) {
+    modifier txExists(uint256 _txIndex) {
         require(_txIndex < transactions.length, "MultiSig: tx does not exist");
         _;
     }
 
-    modifier notExecuted(uint _txIndex) {
+    modifier notExecuted(uint256 _txIndex) {
         require(!transactions[_txIndex].executed, "MultiSig: tx already executed");
         _;
     }
 
-    modifier notConfirmed(uint _txIndex) {
+    modifier notConfirmed(uint256 _txIndex) {
         require(!isConfirmed[_txIndex][msg.sender], "MultiSig: tx already confirmed");
         _;
     }
 
-    modifier notDisabled(uint _txIndex) {
+    modifier notDisabled(uint256 _txIndex) {
         require(_txIndex >= lastDisabledTransactionIndex, "MultiSig: old txs has been disabled");
         _;
     }
 
-    modifier notExpired(uint _txIndex) {
+    modifier notExpired(uint256 _txIndex) {
         require(transactions[_txIndex].expireTimestamp >= block.timestamp || transactions[_txIndex].expireTimestamp == 0, "MultiSig: tx expired");
         _;
     }
@@ -70,7 +70,7 @@ contract MultiSigWallet is IMultiSigWallet {
         _;
     }
 
-    modifier validRequirement(uint ownerCount, uint _required) {
+    modifier validRequirement(uint256 ownerCount, uint256 _required) {
         require(
             ownerCount > 0 && ownerCount <= MAX_OWNER_COUNT && _required <= ownerCount && ownerCount > 1 ? _required > 1 : _required > 0,
             "MultiSig: Invalid requirement"
@@ -85,9 +85,9 @@ contract MultiSigWallet is IMultiSigWallet {
 
     modifier validateSubmitTxInputs(
         address _to,
-        uint _value,
+        uint256 _value,
         bytes memory _data,
-        uint _expireTimestamp
+        uint256 _expireTimestamp
     ) {
         require(_expireTimestamp >= block.timestamp || _expireTimestamp == 0, "already expired");
 
@@ -102,13 +102,17 @@ contract MultiSigWallet is IMultiSigWallet {
         _;
     }
 
-    constructor(address[] memory _owners, uint _numConfirmationsRequired, address _governor) {
+    constructor(
+        address[] memory _owners,
+        uint256 _numConfirmationsRequired,
+        address _governor
+    ) {
         governor = _governor;
         require(_owners.length <= MAX_OWNER_COUNT, "owners limit reached");
         require(_owners.length > 0, "owners required");
         require(_numConfirmationsRequired > 0 && _numConfirmationsRequired <= _owners.length, "invalid number of required confirmations");
 
-        for (uint i = 0; i < _owners.length; i++) {
+        for (uint256 i = 0; i < _owners.length; i++) {
             address owner = _owners[i];
 
             require(owner != address(0), "invalid owner");
@@ -127,7 +131,7 @@ contract MultiSigWallet is IMultiSigWallet {
 
     function removeOwner(address owner) public override onlyWallet ownerExists(owner) {
         isOwner[owner] = false;
-        for (uint i = 0; i < owners.length - 1; i++)
+        for (uint256 i = 0; i < owners.length - 1; i++)
             if (owners[i] == owner) {
                 owners[i] = owners[owners.length - 1];
                 break;
@@ -141,10 +145,13 @@ contract MultiSigWallet is IMultiSigWallet {
         emit OwnerRemoval(owner);
     }
 
-    function addOwners(
-        address[] memory _owners
-    ) public override onlyWallet validRequirement(owners.length + _owners.length, numConfirmationsRequired + _owners.length) {
-        for (uint i = 0; i < _owners.length; i++) {
+    function addOwners(address[] memory _owners)
+        public
+        override
+        onlyWallet
+        validRequirement(owners.length + _owners.length, numConfirmationsRequired + _owners.length)
+    {
+        for (uint256 i = 0; i < _owners.length; i++) {
             address owner = _owners[i];
             require(owner != address(0), "MultiSig: owner address == 0");
             _requireNewOwner(owner);
@@ -157,24 +164,24 @@ contract MultiSigWallet is IMultiSigWallet {
         changeRequirement(numConfirmationsRequired + _owners.length);
     }
 
-    function changeRequirement(uint _required) public override onlyWallet validRequirement(owners.length, _required) {
+    function changeRequirement(uint256 _required) public override onlyWallet validRequirement(owners.length, _required) {
         numConfirmationsRequired = _required;
         emit RequirementChange(_required);
     }
 
     function submitTransaction(
         address _to,
-        uint _value,
+        uint256 _value,
         bytes memory _data,
-        uint _expireTimestamp
+        uint256 _expireTimestamp
     ) public override onlyOwnerOrGov validateSubmitTxInputs(_to, _value, _data, _expireTimestamp) {
-        require(address(this).balance >= _value,"submitTransaction: not enough balance");
-        if(!_to.isContract()){
-            require(_value!= 0, "submitTransaction: value is zero");
-            require(_data.length >0,"submitTransaction: not equal");
+        require(address(this).balance >= _value, "submitTransaction: not enough balance");
+        if (!_to.isContract()) {
+            require(_value != 0, "submitTransaction: value is zero");
+            require(_data.length > 0, "submitTransaction: not equal");
         }
-        uint txIndex = transactions.length;
-        
+        uint256 txIndex = transactions.length;
+
         transactions.push(
             Transaction({ to: _to, value: _value, data: _data, executed: false, numConfirmations: 0, expireTimestamp: _expireTimestamp })
         );
@@ -184,9 +191,16 @@ contract MultiSigWallet is IMultiSigWallet {
         emit SubmitTransaction(txIndex, msg.sender, _to, _value, _data);
     }
 
-    function confirmTransaction(
-        uint _txIndex
-    ) public override onlyOwnerOrGov txExists(_txIndex) notExecuted(_txIndex) notConfirmed(_txIndex) notDisabled(_txIndex) notExpired(_txIndex) {
+    function confirmTransaction(uint256 _txIndex)
+        public
+        override
+        onlyOwnerOrGov
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+        notConfirmed(_txIndex)
+        notDisabled(_txIndex)
+        notExpired(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
 
         _requireTargetCodeNotChanged(transaction.to);
@@ -197,9 +211,15 @@ contract MultiSigWallet is IMultiSigWallet {
         emit ConfirmTransaction(msg.sender, _txIndex);
     }
 
-    function executeTransaction(
-        uint _txIndex
-    ) public override onlyOwnerOrGov txExists(_txIndex) notExecuted(_txIndex) notDisabled(_txIndex) notExpired(_txIndex) {
+    function executeTransaction(uint256 _txIndex)
+        public
+        override
+        onlyOwnerOrGov
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+        notDisabled(_txIndex)
+        notExpired(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
 
         _requireTargetCodeNotChanged(transaction.to);
@@ -218,9 +238,15 @@ contract MultiSigWallet is IMultiSigWallet {
         emit ExecuteTransaction(msg.sender, _txIndex);
     }
 
-    function revokeConfirmation(
-        uint _txIndex
-    ) public override onlyOwnerOrGov txExists(_txIndex) notExecuted(_txIndex) notDisabled(_txIndex) notExpired(_txIndex) {
+    function revokeConfirmation(uint256 _txIndex)
+        public
+        override
+        onlyOwnerOrGov
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+        notDisabled(_txIndex)
+        notExpired(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
 
         require(isConfirmed[_txIndex][msg.sender], "tx not confirmed");
@@ -235,13 +261,23 @@ contract MultiSigWallet is IMultiSigWallet {
         return owners;
     }
 
-    function getTransactionCount() public view override returns (uint) {
+    function getTransactionCount() public view override returns (uint256) {
         return transactions.length;
     }
 
-    function getTransaction(
-        uint _txIndex
-    ) public view override returns (address to, uint value, bytes memory data, bool executed, uint numConfirmations, uint expireTimestamp) {
+    function getTransaction(uint256 _txIndex)
+        public
+        view
+        override
+        returns (
+            address to,
+            uint256 value,
+            bytes memory data,
+            bool executed,
+            uint256 numConfirmations,
+            uint256 expireTimestamp
+        )
+    {
         Transaction memory transaction = transactions[_txIndex];
 
         return (transaction.to, transaction.value, transaction.data, transaction.executed, transaction.numConfirmations, transaction.expireTimestamp);
