+++ contracts/dao/treasury/interfaces/IMultiSigWallet.sol
@@ -4,14 +4,14 @@
 pragma solidity 0.8.13;
 
 interface IMultiSigWallet {
-    event Deposit(address indexed sender, uint amount, uint balance);
-    event SubmitTransaction(uint indexed txIndex, address indexed owner, address indexed to, uint value, bytes data);
-    event ConfirmTransaction(address indexed owner, uint indexed txIndex);
-    event RevokeConfirmation(address indexed owner, uint indexed txIndex);
-    event ExecuteTransaction(address indexed owner, uint indexed txIndex);
+    event Deposit(address indexed sender, uint256 amount, uint256 balance);
+    event SubmitTransaction(uint256 indexed txIndex, address indexed owner, address indexed to, uint256 value, bytes data);
+    event ConfirmTransaction(address indexed owner, uint256 indexed txIndex);
+    event RevokeConfirmation(address indexed owner, uint256 indexed txIndex);
+    event ExecuteTransaction(address indexed owner, uint256 indexed txIndex);
     event OwnerRemoval(address indexed owner);
     event OwnerAddition(address indexed owner);
-    event RequirementChange(uint required);
+    event RequirementChange(uint256 required);
 
     receive() external payable;
 
@@ -19,21 +19,33 @@ interface IMultiSigWallet {
 
     function addOwners(address[] calldata _owners) external;
 
-    function changeRequirement(uint _required) external;
+    function changeRequirement(uint256 _required) external;
 
-    function submitTransaction(address _to, uint _value, bytes memory _data, uint _expireTimestamp) external;
+    function submitTransaction(
+        address _to,
+        uint256 _value,
+        bytes memory _data,
+        uint256 _expireTimestamp
+    ) external;
 
-    function confirmTransaction(uint _txIndex) external;
+    function confirmTransaction(uint256 _txIndex) external;
 
-    function executeTransaction(uint _txIndex) external;
+    function executeTransaction(uint256 _txIndex) external;
 
-    function revokeConfirmation(uint _txIndex) external;
+    function revokeConfirmation(uint256 _txIndex) external;
 
     function getOwners() external returns (address[] memory);
 
-    function getTransactionCount() external returns (uint);
+    function getTransactionCount() external returns (uint256);
 
-    function getTransaction(
-        uint _txIndex
-    ) external returns (address to, uint value, bytes memory data, bool executed, uint numConfirmations, uint expireTimestamp);
+    function getTransaction(uint256 _txIndex)
+        external
+        returns (
+            address to,
+            uint256 value,
+            bytes memory data,
+            bool executed,
+            uint256 numConfirmations,
+            uint256 expireTimestamp
+        );
 }
